Cast day of week to integer

+++ kakaostats.com/view/donor_history.py
@@ -49,7 +49,7 @@ select
     coalesce(points, 0) points,
     coalesce(wus, 0) wus,
     to_char(gd, 'YYYY-MM-DD') as "day",
-    extract(isodow from gd) as dow,
+    extract(isodow from gd)::integer as dow,
     yearweek(gd) as yw
 from (
     select
@@ -78,7 +78,7 @@ from (
     window w as (order by d.data desc)
 ) ss right join (
     select gd::date gd
-    from 
+    from
     generate_series(
         (select data from d_min),
         (select max(data) from datas),
@@ -117,7 +117,7 @@ def donor_history():
     rs = get_rs(queries)
 
     history_json = quote_plus(json.dumps(
-        [[int(d['points']), d['wus'], d['day'], int(d['dow'])] for d in rs['daily'][:-1]]
+        [[int(d['points']), d['wus'], d['day'], d['dow']] for d in rs['daily'][:-1]]
         , separators=(',',':')))
 
     return render_template(
